requirements.txt
numpy
pandas
matplotlib
seaborn
plotly

jupyter

pymongo

imblearn
catboost
scikit-learn
xgboost

python-dotenv
from_root
dill
PyYAML
neuro_mf
fastapi
uvicorn
jinja2
python-multipart
python-box

mlflow
dvc
dagshub
